+++ CHANGELOG.md
@@ -1,9 +1,16 @@
-## [0.0.1] - 02/01/2020 (January 2020)
+#### [0.0.1] - 02/01/2020 (January 2020)
 
 - Added DateIndicator() widget that display horizontal list with circle date holder.
 
-## [0.0.2] - 13/01/2020 (January 2020)
+#### [0.0.2] - 13/01/2020 (January 2020)
 
 - Provided a way to add custom colors to indicator, date holder and texts.
 - onHolderTap function to get selected day.
 - Example app added.
+
+#### [0.0.3] - 14/01/2020 (January 2020)
+
+- Provided a list parameter to add available bubble under date holder.
+- Add another parameter to add initial selected day.
+- Example app updated with latest changes.
+- README updated with new example.